More changes for model format to use msgpack.

+++ src/include/learning/perceptron/score.h
@@ -91,7 +91,7 @@ operator >>(std::istream &is, CScore<int> &score) {
   assert(nitems == 2); (void)nitems;
   score[0] = mp::read_int(is);
   score[1] = mp::read_int(is);
-  return is ;
+  return is;
 }
 
 template <>
@@ -101,7 +101,7 @@ operator >>(std::istream &is, CScore<long> &score) {
   assert(nitems == 2); (void)nitems;
   score[0] = mp::read_int(is);
   score[1] = mp::read_int(is);
-  return is ;
+  return is;
 }
 
 template <>
@@ -111,7 +111,7 @@ operator >>(std::istream &is, CScore<float> &score) {
   assert(nitems == 2); (void)nitems;
   score[0] = mp::read_int(is);
   score[1] = mp::read_int(is);
-  return is ;
+  return is;
 }
 
 
@@ -132,8 +132,8 @@ template <>
 inline std::ostream &
 operator <<(std::ostream &os, const CScore<long> &score) {
   mp::write_array_size(os, 2);
-  mp::write_int(os, static_cast<int>(score[0]));
-  mp::write_int(os, static_cast<int>(score[1]));
+  mp::write_int(os, score[0]);
+  mp::write_int(os, score[1]);
   return os;
 }
